+++ Assets/Scripts/FirstPersonController.cs
@@ -295,7 +295,7 @@ public class FirstPersonController : MonoBehaviour
             ball.TryKick(gameObject);
     }
 
-    private const string GameVersion = "0.3.70-dev";
+    private const string GameVersion = "0.3.71-dev";
 
     private float lastSpeed;
     private bool lastSprinting;